requirements.txt
flask
line-bot-sdk
python-dotenv
torch
torchvision
Pillow
waitress
